backend/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
pymongo==4.5.0
PyJWT==2.8.0
werkzeug==2.3.7
python-dotenv==1.0.0
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.0
numpy==1.24.3
pandas==1.5.3
openai==1.3.0
reportlab==4.0.4
jinja2==3.1.2
